# Conflicts:
#	src/main/java/com/musu/repository/OrdersRepository.java

+++ src/main/java/com/musu/service/UserService.java
@@ -7,5 +7,5 @@ public interface UserService {
     void save(User user);
     List<User> findAll();
     User findByUsername(String username);
-    void deleteUser(Long id);
+    void deleteUser(int id);
 }
